requirements.txt
pandas
openai
faiss-cpu
numpy
langchain
PyMuPDF
flask
langchain_openai